functions/product_image_analyzer/requirements.txt
functions-framework==3.*
google-genai==1.*
requests==2.*
supabase==2.*
